Rebuild animator state machines and fix movement
Fixed movement and its animations as well as firing animations in a new 3x6 state machine format instead of a webbed mess of 18 states. The engine flame has its own object with the ship as its parent and its own animator with 3 states. The ship's animator now has only 6 states to take care of. Animations with 2 identical frames were used to make the illusion of images of the non-firing ship.